+++ plugins/co/Text/TextObject.cpp
@@ -204,7 +204,7 @@ void TextObject::setSettings (Setting &set)
 {
   QString s = set.getData("Date");
   date.setDate(s);
-  value = set.getFloat("Value");
+  value = set.getDouble("Value");
   color.setNamedColor(set.getData("Color"));
   plot = set.getData("Plot");
   name = set.getData("Name");
